+++ dom/src/main/java/react4j/dom/ReactDOM.java
@@ -7,9 +7,12 @@ import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
+import org.realityforge.braincheck.BrainCheckConfig;
 import react4j.core.Procedure;
+import react4j.core.React;
 import react4j.core.ReactNode;
 import react4j.dom.proptypes.html.HtmlGlobalFields;
+import static org.realityforge.braincheck.Guards.*;
 
 /**
  * Core interface into React DOM library.
@@ -52,9 +55,30 @@ public class ReactDOM
    * @return a reference to the created React Component, DOM Node, Portal or null (stateless components).
    */
   @Nullable
-  public static native Object render( @Nonnull ReactNode node,
-                                      @Nonnull Element container,
-                                      @Nullable Procedure onUpdate );
+  @JsOverlay
+  public static Object render( @Nonnull final ReactNode node,
+                               @Nonnull final Element container,
+                               @Nullable final Procedure onUpdate )
+  {
+    if ( BrainCheckConfig.checkInvariants() )
+    {
+      invariant( () -> React.isValidElement( node ), () -> "ReactDOM.render passed a non ReactElement" );
+    }
+    return render0( node, container, onUpdate );
+  }
+
+  /**
+   * Internal native method for rendering component./
+   *
+   * @param node      the react node to render.
+   * @param container the DOM element to render into.
+   * @param onUpdate  the callback invoked when rendering is complete.
+   * @return a reference to the created React Component, DOM Node, Portal or null (stateless components).
+   */
+  @JsMethod( name = "render" )
+  private static native Object render0( @Nonnull ReactNode node,
+                                        @Nonnull Element container,
+                                        @Nullable Procedure onUpdate );
 
   /**
    * Render a React element into the DOM in the supplied container.